moves sync destinations types out of app-types
